requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0

# Web interface
streamlit>=1.28.0
plotly>=5.17.0

# Configuration and utilities
pyyaml>=6.0
pathlib2>=2.3.7

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For enhanced visualization
seaborn>=0.12.0
imageio>=2.31.0

# Optional: For advanced features
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
